% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_funcs.R
\name{create_NEW}
\alias{create_NEW}
\title{Create new student info files, one per class}
\usage{
create_NEW(path = "~/Documents/Students", pattern = "Fall_2018")
}
\arguments{
\item{path}{folder where student files are stored}

\item{pattern}{pattern to match in filenames}
}
\description{
Clears previous NEW files, creates new ones from downloaded CourseWorks files, and writes to the path specified, prefixing the original filename with "NEW". Nothing is returned.

Call only if new files have been downloaded from CourseWorks
}
\examples{
create_NEW()
create_NEW("~/Downloads", "Spring_2019")
}
