% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_funcs.R
\name{get_student_data}
\alias{get_student_data}
\title{Get student info in tidy format}
\usage{
get_student_data(path = "~/Documents/Students", pattern = "NEW")
}
\arguments{
\item{path}{folder where student files are stored}

\item{pattern}{pattern to match in filenames}
}
\value{
a tidy data frame with student info
}
\description{
Reads in all files with NEW prefix and combines into into one tidy \code{tibble}.
}
\examples{
stud_df <- get_student_data()
stud_df <- get_student_data("~/Downloads")
}
