% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_funcs.R
\name{hmk}
\alias{hmk}
\title{Get homework score}
\usage{
hmk(name, num, data)
}
\arguments{
\item{name}{string containing all or part of student name to search for}

\item{num}{homework number}

\item{data}{data frame of student info in tidy format}
}
\value{
a tibble with columns: Student, Homework, HmkScore, Section
}
\description{
Returns homework score(s) for any matches in student name and assignment number.
}
\examples{
stud_df <- get_student_data("~/Downloads")
hmk("Sarah", 1, stud_df)
}
