% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_funcs.R
\name{lpr}
\alias{lpr}
\title{Get number of late passes remaining}
\usage{
lpr(name, data)
}
\arguments{
\item{name}{string containing all or part of student name to search for}

\item{data}{data frame of student info in tidy format}
}
\value{
a tibble with columns: Student, Late Passes Remaining, Section
}
\description{
Returns number of late passes remaining (only applies to 1201)
}
\examples{
stud_df <- get_student_data("~/Downloads")
lpr("Jing", stud_df)
}
