% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_funcs.R
\name{missed_test}
\alias{missed_test}
\title{List students who missed tests}
\usage{
missed_test(num = 1, data = NULL)
}
\arguments{
\item{num}{test number}

\item{data}{database of student info}
}
\value{
a tibble with columns: Student, Section
}
\description{
Returns students with NA for a particular test number
}
\examples{
stud_df <- get_student_data("~/Downloads")
missed_test(1, stud_df)
}
