% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_funcs.R
\name{name2uni}
\alias{name2uni}
\title{What's the UNI?}
\usage{
name2uni(name, data = NULL)
}
\arguments{
\item{name}{string containing all or part of student name to search for}

\item{data}{data frame of student info in tidy format}
}
\value{
a tibble with columns: Student, UNI, Section
}
\description{
Returns UNI(s) for any matches in student name
}
\examples{
stud_df <- get_student_data("~/Downloads")
name2uni("Daisy", stud_df)
}
