% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_funcs.R
\name{tst}
\alias{tst}
\title{Get test score}
\usage{
tst(name, data = NULL)
}
\arguments{
\item{name}{string containing all or part of student name to search for}

\item{data}{data frame of student info in tidy format}
}
\value{
a tibble with columns: Student, Test, TestScore, Section
}
\description{
Returns test score(s) for any matches in student name.
}
\examples{
stud_df <- get_student_data("~/Downloads")
tst("Emily")
}
