% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_funcs.R
\name{uni2name}
\alias{uni2name}
\title{What's the name?}
\usage{
uni2name(uni, data = NULL)
}
\arguments{
\item{uni}{string containing all or part of student UNI to search for}

\item{data}{data frame of student info in tidy format}
}
\value{
a tibble with columns: Student, UNI, Section
}
\description{
Returns student name(s) for any match in student UNI
}
\examples{
stud_df <- get_student_data("~/Downloads")
uni2name("jtr13")
}
