% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_funcs.R
\name{wc}
\alias{wc}
\title{Which class?}
\usage{
wc(name, data = NULL)
}
\arguments{
\item{name}{string containing all or part of student name to search for}

\item{data}{data frame of student info in tidy format}
}
\value{
a tibble with columns: Student, Section
}
\description{
Returns class name for any matches in student name
}
