% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submissions.R
\name{zip_to_drive}
\alias{zip_to_drive}
\title{Takes a downloaded zip of submissions from CourseWorks and uploads sheet of text entry information to Google Drive}
\usage{
zip_to_drive(path = "~/Downloads", pattern = "test")
}
\arguments{
\item{path}{folder where student files are stored}

\item{pattern}{pattern to match in filenames}
}
\description{
Gets group number (or name) from file name, title from h1 tag, and content from p tags. Name of Google Sheet is determined by zip file name.

Call only when creating a new sheet on GoogleDrive.
}
\examples{
nss <- zip_to_drive("~/Downloads", "Final_Project_Groups.zip")
drive_browse(nss)
}
